some fixes in burger entity tests
