+++ Restaraunt.Application/Products/Drinks/Commands/CreateDrink/CreateDrinkCommand.cs
@@ -3,7 +3,7 @@
 namespace Restaraunt.Application.Products.Drinks.Commands.CreateDrink
 {
 	public sealed record CreateDrinkCommand
-		(Guid Id,
+		(
 		string Name,
 		string Description,
 		double Price,
